- Split single POM file into multiple page-specific files for better structure and maintainability - Improved test functions for better reuse - Ensured all locators are stored in named variables for consistency and readability
