+++ package.json
@@ -9,7 +9,8 @@
   "description": "",
   "devDependencies": {
     "@playwright/test": "^1.52.0",
-    "@types/node": "^22.15.29"
+    "@types/node": "^22.15.29",
+    "globals": "^16.2.0"
   },
   "dependencies": {
     "dotenv": "^16.5.0"
